most of some gets and post update

+++ src/context/authContext.js
@@ -7,23 +7,39 @@ export const AuthContextProvider = ({ children }) => {
   const [currentUser, setCurrentUser] = useState(JSON.parse(localStorage.getItem("userdata")) || null);
 
   const login = async (inputs) => {
-    // const res = await api.post("/login", inputs, {
-    //   withCredentials: true,
-    // });
-    const response = await api.post("/login", inputs);
+    const response = await api.post("/auth/login", inputs);
+
     setCurrentUser(response.data);
+
+    return response;
   };
 
-  const logout = async () => {
-    // Define your token
+  const register = async (inputs) => {
+    const response = await api.post("/auth/register", inputs);
+
+    return response;
+  };
+
+  const me = async () => {
     const token = currentUser.token;
 
-    // Define headers with Authorization Bearer token
     const headers = {
       Authorization: `Bearer ${token}`,
     };
 
-    const response = await api.delete("/logout", { headers });
+    const response = await api.get("/auth/me", { headers });
+
+    return response;
+  };
+
+  const logout = async () => {
+    const token = currentUser.token;
+
+    const headers = {
+      Authorization: `Bearer ${token}`,
+    };
+
+    const response = await api.delete("/auth/logout", { headers });
     setCurrentUser(null);
   };
 
@@ -31,5 +47,5 @@ export const AuthContextProvider = ({ children }) => {
     localStorage.setItem("userdata", JSON.stringify(currentUser));
   }, [currentUser]);
 
-  return <AuthContext.Provider value={{ currentUser, login, logout }}>{children}</AuthContext.Provider>;
+  return <AuthContext.Provider value={{ currentUser, login, register, me, logout }}>{children}</AuthContext.Provider>;
 };
